+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-- NA
+- Added [Convert-CharCase][] the corresponding delimiter with the -AsSentence switch.
 
 ## [0.8.0] - 2025-04-27